+++ src/selector/focusable.supports.js
@@ -2,7 +2,6 @@
 import memorizeResult from '../supports/memorize-result';
 import canFocusAudioWithoutControls from '../supports/focus-audio-without-controls';
 import canFocusVideoWithoutControls from '../supports/focus-video-without-controls';
-import canFocusSvgMethod from '../supports/svg-focus-method';
 import canFocusTable from '../supports/focus-table';
 import canFocusFieldset from '../supports/focus-fieldset';
 import canFocusSummary from '../supports/focus-summary';
@@ -12,7 +11,6 @@ export default memorizeResult(function() {
   return {
     canFocusAudioWithoutControls: canFocusAudioWithoutControls(),
     canFocusVideoWithoutControls: canFocusVideoWithoutControls(),
-    canFocusSvgMethod: canFocusSvgMethod(),
     canFocusTable: canFocusTable(),
     canFocusFieldset: canFocusFieldset(),
     canFocusSummary: canFocusSummary(),