+++ test/unit/query.focusable.test.js
@@ -40,7 +40,35 @@ define([
           '#image-map-area',
           supports.canFocusObjectSvg && '#object-svg',
           supports.canFocusObjectSvg && '#object-tabindex-svg',
-          supports.canFocusSvgMethod && '#svg-link',
+          supports.svgFocusMethod && '#svg-link',
+          supports.canFocusAudioWithoutControls && '#audio',
+          '#audio-controls',
+          '#input',
+          '#input-tabindex--1',
+          '#span-contenteditable',
+          '#img-ismap-link',
+          '#focusable-flexbox',
+        ].filter(Boolean);
+
+        expect(result).to.deep.equal(expected);
+      },
+
+      includeOnlyTabbable: function() {
+        var result = queryFocusable({
+          includeOnlyTabbable: true,
+        }).map(fixture.nodeToString);
+
+        var expected = [
+          '#tabindex--1',
+          '#tabindex-0',
+          '#tabindex-1',
+          supports.canFocusInvalidTabindex && '#tabindex-bad',
+          '#link',
+          '#link-tabindex--1',
+          '#image-map-area',
+          supports.canFocusObjectSvg && '#object-svg',
+          supports.canFocusObjectSvg && '#object-tabindex-svg',
+          '#svg-link',
           supports.canFocusAudioWithoutControls && '#audio',
           '#audio-controls',
           '#input',
@@ -136,7 +164,7 @@ define([
           '#image-map-area',
           supports.canFocusObjectSvg && '#object-svg',
           supports.canFocusObjectSvg && '#object-tabindex-svg',
-          supports.canFocusSvgMethod && '#svg-link',
+          supports.svgFocusMethod && '#svg-link',
           supports.canFocusAudioWithoutControls && '#audio',
           '#audio-controls',
           '#input',